refactor: simplify chapter internals
