+++ backend/app/alembic/versions/3a1b9d7f6c20_add_album_chapters.py
@@ -1,4 +1,4 @@
-"""Add album chapters.
+"""Move album content into chapters.
 
 Revision ID: 3a1b9d7f6c20
 Revises: 1f6d2e8c9a0b
@@ -20,7 +20,15 @@ def upgrade() -> None:
         "album",
         sa.Column("chapters", sa.JSON(), nullable=False, server_default="[]"),
     )
+    op.drop_column("album", "title")
+    op.drop_column("album", "subtitle")
+    op.drop_column("album", "front_cover_photo")
+    op.drop_column("album", "back_cover_photo")
 
 
 def downgrade() -> None:
+    op.add_column("album", sa.Column("back_cover_photo", sa.String(255)))
+    op.add_column("album", sa.Column("front_cover_photo", sa.String(255)))
+    op.add_column("album", sa.Column("subtitle", sa.String(255)))
+    op.add_column("album", sa.Column("title", sa.String(255)))
     op.drop_column("album", "chapters")
